# rs-processing-common/request_parking_lot/config/openapi/s1pro-api_openapi.yaml
...
openapi: 3.0.1
info:
  title: S1PRO-API
  description: >
    This API provides acces to the S1PRO subsystem
      
      ## Changelog
      ### V1.0.0
      - initial version
      
      ## Usage Notes
      
      ### API security
      The API is protected by a API key. This is a token (string) agreed by server and client and used to authenticate the client. The client needs to provide the (configurable) API key on every request as a HTTP-Header attribute, e.g. `ApiKey: <yourApiKey>`. The server validates the key send by the client and only allows access if the key is known to the server, else a HTTP error 403 is raised.
      
      ### View and manage failed processings 
      To get a list of failed processings call `GET /failedProcessings`. It will return a list of all failed processings. Operators can review the entries and either restart the processing by `POST /failedProcessings/{id}/restart`, resubmit a failed processing by `POST /failedProcessings/{id}/resubmit` or delete a failed processing by `DELETE /failedProcessings/{id}`. In either case the failed processing is no longer present in the list of failed processings. In case of a restart a failed processing is issued again to the queues eventually resulting in a new processing. Whereas in the case of a resubmit the message that preceded the failed processing is reissued.
      
      A single failed procesing can be viewed by `GET /failedProcessings/{id}`.
      
...
      
...
      
...
      
...
      
...
      
...
      
...
      
...
servers:
...
paths:
  /failedProcessings:
    get:
      tags:
      - FailedProcessings
      summary: get the list of failed processings ordered by creation time (ascending)
      operationId: getFailedProcessingList
      parameters:
      - name: ApiKey
        in: header
        required: true
        style: simple
        schema:
          type: string
      responses:
        200:
          description: The request succeeded
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FailedProcessing'
        400:
          description: Invalid parameters supplied
          content: {}
        403:
          description: Forbidden, e.g. Invalid API key supplied
          content: {}
        500:
          description: Internal Server Error
          content: {}
      deprecated: false
      security:
      - ApiKey: []
  /failedProcessings/count:
    get:
      tags:
      - FailedProcessings
      summary: get the number of failed processings
      operationId: countFailedProcessingList
      parameters:
      - name: ApiKey
        in: header
        required: true
        style: simple
        schema:
          type: string
      responses:
        200:
          description: The request succeeded
          content:
            application/json:
              schema:
                type: integer
                format: int64
                description: number of failed processings
        400:
          description: Invalid parameters supplied
          content: {}
        403:
          description: Forbidden, e.g. Invalid API key supplied
          content: {}
        500:
          description: Internal Server Error
          content: {}
      deprecated: false
      security:
      - ApiKey: []
  /failedProcessings/{id}:
    get:
      tags:
      - FailedProcessings
      summary: get failed processing by id
      operationId: getFailedProcessingById
      parameters:
      - name: ApiKey
        in: header
        required: true
        style: simple
        schema:
          type: string
      - name: id
        in: path
        description: "`id` of FailedProcessing to get"
        required: true
        style: simple
        schema:
          type: string
      responses:
        200:
          description: The request succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedProcessing'
        400:
          description: Invalid parameters supplied
          content: {}
        403:
          description: Forbidden, e.g. Invalid API key supplied
          content: {}
        404:
          description: Item not found
          content: {}
        500:
          description: Internal Server Error
          content: {}
      deprecated: false
      security:
      - ApiKey: []
    delete:
      tags:
      - FailedProcessings
      summary: Deletes a failed processing by id
      operationId: deleteFailedProcessing
      parameters:
      - name: ApiKey
        in: header
        required: true
        style: simple
        schema:
          type: string
      - name: id
        in: path
        description: "`id` of FailedProcessing to delete"
        required: true
        style: simple
        schema:
          type: string
      responses:
        200:
          description: The request succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        400:
          description: Invalid parameters supplied
          content: {}
        403:
          description: Forbidden, e.g. Invalid API key supplied
          content: {}
        404:
          description: Item not found
          content: {}
        500:
          description: Internal Server Error
          content: {}
      security:
      - ApiKey: []
  /failedProcessings/delete:
    post:
      tags:
      - FailedProcessings
      summary: delete multiple failed processings in a batch. Ids for not existing failed processings are ignored.
      operationId: deleteFailedProcessingsInBatch
      parameters:
      - name: ApiKey
        in: header
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  description: "array of `id`s of FailedProcessing to delete"
                  type: array
                  items:
                    type: string
      responses:
        200:
          description: The request succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        400:
          description: Invalid parameters supplied
          content: {}
        403:
          description: Forbidden, e.g. Invalid API key supplied
          content: {}
        404:
          description: Item not found
          content: {}
        500:
          description: Internal Server Error
          content: {}
      deprecated: false
      security:
      - ApiKey: []
  /failedProcessings/{id}/restart:
    post:
      tags:
      - FailedProcessings
      summary: restart a failed processing job. This removes the failed processing from
        the list and restarts it.
      operationId: restartFailedProcessing
      parameters:
      - name: ApiKey
        in: header
        required: true
        style: simple
        schema:
          type: string
      - name: id
        in: path
        description: "`id` of FailedProcessing job to restart"
        required: true
        style: simple
        schema:
          type: string
      responses:
        200:
          description: The request succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        400:
          description: Invalid parameters supplied
          content: {}
        403:
          description: Forbidden, e.g. Invalid API key supplied
          content: {}
        404:
          description: Item not found or no restart available
          content: {}
        500:
          description: Internal Server Error
          content: {}
      deprecated: false
      security:
      - ApiKey: []
  /failedProcessings/restart:
    post:
      tags:
      - FailedProcessings
      summary: restart multiple failed processing jobs in a batch. This removes the failed processings from
        the list and restarts them. Ids for not existing or not restartable failed processings are ignored.
      operationId: restartFailedProcessingsInBatch
      parameters:
      - name: ApiKey
        in: header
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  description: "array of `id`s of FailedProcessing jobs to restart"
                  type: array
                  items:
                    type: string
      responses:
        200:
          description: The request succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        400:
          description: Invalid parameters supplied
          content: {}
        403:
          description: Forbidden, e.g. Invalid API key supplied
          content: {}
        404:
          description: Item not found
          content: {}
        500:
          description: Internal Server Error
          content: {}
      deprecated: false
  /failedProcessings/{id}/resubmit:
    post:
      tags:
      - FailedProcessings
      summary: resubmit a failed processing job. This removes the failed processing from
        the list and resubmits it.
      operationId: resubmitFailedProcessing
      parameters:
      - name: ApiKey
        in: header
        required: true
        style: simple
        schema:
          type: string
      - name: id
        in: path
        description: "`id` of FailedProcessing job to resubmit"
        required: true
        style: simple
        schema:
          type: string
      responses:
        200:
          description: The request succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        400:
          description: Invalid parameters supplied
          content: {}
        403:
          description: Forbidden, e.g. Invalid API key supplied
          content: {}
        404:
          description: Item not found or no resubmit available
          content: {}
        500:
          description: Internal Server Error
          content: {}
      deprecated: false
      security:
      - ApiKey: []
  /failedProcessings/resubmit:
    post:
      tags:
      - FailedProcessings
      summary:  multiple failed processing jobs in a batch. This removes the failed processings from
        the list and resubmits them. Ids for not existing or not resubmittable failed processings are ignored.
      operationId: resubmitFailedProcessingsInBatch
      parameters:
      - name: ApiKey
        in: header
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  description: "array of `id`s of FailedProcessing jobs to resubmit"
                  type: array
                  items:
                    type: string
      responses:
        200:
          description: The request succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        400:
          description: Invalid parameters supplied
          content: {}
        403:
          description: Forbidden, e.g. Invalid API key supplied
          content: {}
        404:
          description: Item not found
          content: {}
        500:
          description: Internal Server Error
          content: {}
      deprecated: false
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        entity:
          type: string
          enum:
           - FailedProcessing
        action:
          type: string
          enum:
           - restart
           - resubmit
           - delete
          description: action performed
        idsWithSuccess:
          type: array
          description: list of ids, where 'action' was be applied succesfully
          items:
            type: string
        idsSkipped:
          type: array
          description: list of ids, where 'action' was not applied, e.g. object not found or other reasons
          items:
            type: string
      description: Message container in case of successfull operations
...
    FailedProcessing:
      type: object
      properties:
        id:
          type: string
          description: id of the failed processing
        topic:
          type: string
          description: kafka topic of the failed processing
        missionId:
          type: string
          description: sentinel mission id
        errorLevel:
          type: string
          enum:
            - ERROR
            - WARNING
            - NOT_DEFINED
        failureDate:
          type: string
          description: date when the processing failed
          format: date-time
        failureMessage:
          type: string
          description: error message explaining the failure
        stacktrace:
          type: string
          description: application stack trace
        message:
          #processingDetails is very generic and there probably will be some refinement in later versions of this API to be more precise on potential content
          type: object
          description: "`dto` object from kafka message"
          oneOf:
            - $ref: '#/components/schemas/IngestionEventDto'
            - $ref: '#/components/schemas/CatalogEventDto'
            - $ref: '#/components/schemas/ProductionEventDto'
            - $ref: '#/components/schemas/CompressionEventDto'
            - $ref: '#/components/schemas/IngestionJobDto'
            - $ref: '#/components/schemas/CatalogJobDto'
            - $ref: '#/components/schemas/IpfPreparationJobDto'
            - $ref: '#/components/schemas/IpfExecutionJobDto'
            - $ref: '#/components/schemas/CompressionJobDto'
            - $ref: '#/components/schemas/PripJobDto'
        retryCounter:
          type: integer
          format: int64
          description: number of times that processing has been restarted
    AbstractDto:
      description: base object for all DTO objects on failedProcessings
      type: object
      properties:
        uid: 
          type: string
          format: uuid
          description: event or job id
          example: ''
        keyObjectStorage:
          type: string
          description: usually the product name, but can also be e.g. file name
          example: ''
        productFamily:
          type: string
          deprecated: true
          example: ''
        creationDate:
          type: string
          description: creation date of the job
          example: '2020-01-20T12:40:37.464424Z'
        allowedActions:
          type: array
          items:
            type: string
            enum:
              - RESTART
              - RESUBMIT
              - NO_ACTION
          example: [RESTART]
        debug:
          type: boolean
          default: false
          description: flag for operator debug mode
          example: true
        demandType:
          type: string
          enum:
            - EXTERNAL_DEMAND
            - EXTERNAL_CUSTOM_DEMAND
            - INTERNAL_DEMAND
            - SYSTEMATIC
          description: demand type is to indicate if a processing was triggered from an user, an operator or systematically
          example: 'EXTERNAL_DEMAND'
        retryCounter:
          type: integer
          description: retry counter for explicit restarts via REST
          example: 1
    AbstractEventDto:
      description: data transfer object for events
      allOf: 
        - $ref: '#/components/schemas/AbstractDto'
    IngestionEventDto:
      description: data transfer object for ingestion events
      allOf: 
        - $ref: '#/components/schemas/AbstractEventDto'
        - type: object
          properties:
            relativePath:
              type: string
              description: relative path from base url to data
              example: ''
            productName:
              type: string
              description: product name
              example: 'S1A_WV_RAW__0SSV_20200120T124037_20200120T124728_030884_038B5D_3C86.SAFE'
            productSizeByte:
              type: integer
              format: int64
              description: size of the product
              example: 951324687
            stationName:
              type: string
              description: the station name
              example: 'WILE'
            mode:
              type: string
              description: processing mode
              example: ''
            timeliness:
              type: string
              description: timeliness
              example: 'FAST24'
    CatalogEventDto:
      description: data transfer object for catalog events
      allOf: 
        - $ref: '#/components/schemas/AbstractEventDto'
        - type: object
          properties:
            productName:
              type: string
              description: product name
              example: 'S1A_WV_RAW__0SSV_20200120T124037_20200120T124728_030884_038B5D_3C86.SAFE'
            product_type:
              type: string
              description: the product type
              example: ''
            metadata:
              type: object
              additionalProperties: true
    ProductionEventDto:
      description: data transfer object for production events
      allOf: 
        - $ref: '#/components/schemas/AbstractEventDto'
        - type: object
          properties:
            productName:
              type: string
              description: product name
              example: 'S1A_WV_RAW__0SSV_20200120T124037_20200120T124728_030884_038B5D_3C86.SAFE'
            mode:
              type: string
              description: processing mode
              example: ''
            oqcFlag:
              type: string
              description: quality check status
              enum:
                - CHECKED_OK
                - CHECKED_NOK
                - NOT_CHECKED
            timeliness:
              type: string
              description: timeliness
              example: 'FAST24'
    CompressionEventDto:
      description: data transfer object for compression events
      allOf: 
        - $ref: '#/components/schemas/AbstractEventDto'
    AbstractJobDto:
      description: data transfer object for jobs
      allOf: 
        - $ref: '#/components/schemas/AbstractDto'
    IngestionJobDto:
      description: data transfer object for ingestion jobs
      allOf: 
        - $ref: '#/components/schemas/AbstractJobDto'
        - type: object
          properties:
            pickupBaseURL:
              type: string
              description: base URL for the pickup point
              example: ''
            relativePath:
              type: string
              description: relative path from base url to data
              example: ''
            productName:
              type: string
              description: product name
              example: 'S1A_WV_RAW__0SSV_20200120T124037_20200120T124728_030884_038B5D_3C86.SAFE'
            productSizeByte:
              type: integer
              format: int64
              description: size of the product
              example: 951324687
            stationName:
              type: string
              description: the station name
              example: 'WILE'
            mode:
              type: string
              description: processing mode
              example: ''
            timeliness:
              type: string
              description: timeliness
              example: 'FAST24'
    CatalogJobDto:
      description: data transfer object for catalog jobs
      allOf: 
        - $ref: '#/components/schemas/AbstractJobDto'
        - type: object
          properties:
            productName:
              type: string
              description: product name
              example: 'S1A_WV_RAW__0SSV_20200120T124037_20200120T124728_030884_038B5D_3C86.SAFE'
            relativePath:
              type: string
              description: relative path
              example: ''
            stationName:
              type: string
              description: the station name
              example: 'WILE'
            mode:
              type: string
              description: processing mode
              example: ''
            oqcFlag:
              type: string
              description: quality check status
              enum:
                - CHECKED_OK
                - CHECKED_NOK
                - NOT_CHECKED
            timeliness:
              type: string
              description: timeliness
              example: 'FAST24'
...
    IpfExecutionJobDto:
      description: data transfer object for production execution jobs
      allOf: 
        - $ref: '#/components/schemas/AbstractJobDto'
        - type: object
          properties:
            productProcessMode:
              type: string
              description: product process mode
              example: ''
            workDirectory:
              type: string
              description: local work directory
              example: ''
            jobOrder:
              type: string
              description: local path to the job order
              example: ''
            timeliness:
              type: string
              description: timeliness
              example: 'FAST24'
            inputs:
              type: array
              items:
                type: object
                properties:
                  family:
                    type: string
                    description: product family
                    example: ''
                  localPath:
                    type: string
                    description: local path on the target host
                    example: ''
                  contentRef:
                    type: string
                    description: The reference to the content. Can be the object storage or directly in the string according the family
                    example: ''
            outputs:
              type: array
              items:
                type: object
                properties:
                  family:
                    type: string
                    description: product family
                    example: ''
                  regexp:
                    type: string
                    description: regular expression
                    example: ''
                  oqcCheck:
                    type: boolean
                    description: flag if oqcCheck for the wrapper required for this output
                    example: ''
            pools:
              type: array
              items:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      type: object
                      properties:
                        binaryPath:
                          type: string
                          description: absolute path of the binary
                          example: '' 
    CompressionJobDto:
      description: data transfer object for compressions
      allOf: 
        - $ref: '#/components/schemas/AbstractJobDto'
        - type: object
          properties:
            outputProductFamily:
              type: string
              description: product family for output
              example: ''
            outputKeyObjectStorage:
              type: string
              description: object store key for output product
              example: ''
            compressionDirection:
              type: string
              description: flag to indicate compression or decompression task
              enum:
                - UNDEFINED
                - COMPRESS
                - UNCOMPRESS
              example: COMPRESS
    PripJobDto:
      description: data transfer object for PRIP jobs
      allOf: 
        - $ref: '#/components/schemas/AbstractJobDto'
        - type: object
          properties:
            evictionDate:
              type: string
              description: eviction time
              example: '2021-01-20T12:40:37.464424Z'
...
  securitySchemes:
    ApiKey:
      type: apiKey
      name: ApiKey
      in: header